requirements.txt
torch
transformers
accelerate
bitsandbytes
vosk
piper-tts
sounddevice
soundfile
numpy
scipy
